fix: update to virtool-workflow 5.1.0
BREAKING CHANGE: Only supports API versions 12.0.0+.
