+++ Dockerfile
@@ -1,21 +1,20 @@
-FROM rust:1.60.0-slim-buster as rust
+FROM rust:1.65-slim-buster as rust
 WORKDIR /build
 COPY /utils/count_nucleotides_and_seqs/ /build/
 RUN ls
 RUN cargo build -r
 
-FROM virtool/workflow:4.2.1 as test
-WORKDIR /test
-COPY poetry.lock pyproject.toml /test/
-RUN pip install poetry
-RUN poetry install
-COPY --from=rust /build/target/release/count_nucleotides_and_seqs /test/
-COPY tests / test/
-COPY workflow.py /test/
-RUN poetry run pytest
-
-
-FROM virtool/workflow:4.2.1
+FROM virtool/workflow:5.1.0 as build
 WORKDIR /workflow
 COPY --from=rust /build/target/release/count_nucleotides_and_seqs /workflow/
-COPY workflow.py /workflow/workflow.py
+COPY fixtures.py workflow.py /workflow/
+
+FROM virtool/workflow:5.1.0 as test
+WORKDIR /test
+COPY poetry.lock pyproject.toml /test/
+RUN curl -sSL https://install.python-poetry.org | python -
+RUN poetry install --with dev
+COPY --from=rust /build/target/release/count_nucleotides_and_seqs /test/
+COPY tests / test/
+COPY fixtures.py workflow.py pytest.ini /test/
+RUN poetry run pytest
