+++ pyproject.toml
@@ -5,13 +5,15 @@ description = ""
 authors = ["Ian Boyes <igboyes@gmail.com>"]
 
 [tool.poetry.dependencies]
-python = "^3.8"
-virtool-workflow = "^4.2.1"
+python = "^3.10"
+virtool-workflow = "^5.0.0"
+pyfixtures = "^1.1.0"
 
-[tool.poetry.dev-dependencies]
-pytest-datafiles = "^2.0.1"
-pytest-aiohttp = "^1.0.4"
+[tool.poetry.group.dev.dependencies]
 pytest = "^7.1.2"
+pytest-aiohttp = "^1.0.4"
+pytest-datafiles = "^2.0.1"
+pytest-asyncio = "^0.17.0"
 
 [build-system]
 requires = ["poetry-core>=1.0.0"]
